+++ lib/commercetools-api/src/Models/TaxCategory/TaxRate.php
@@ -21,6 +21,7 @@ interface TaxRate extends JsonObject
     public const FIELD_COUNTRY = 'country';
     public const FIELD_STATE = 'state';
     public const FIELD_SUB_RATES = 'subRates';
+    public const FIELD_TAX_ROUNDING_TARGET = 'taxRoundingTarget';
 
     /**
      * <p>Present if the TaxRate is part of a <a href="ctp:api:type:TaxCategory">TaxCategory</a>.
@@ -92,6 +93,15 @@ interface TaxRate extends JsonObject
     public function getSubRates();
 
     /**
+     * <p>Determines which of the net price and the tax amount the <code>taxRoundingMode</code> of the Cart or Order is applied to, when this TaxRate is included in the price. Ignored if <code>includedInPrice</code> is <code>false</code>.</p>
+     * <p>Always returned by the API. Can be omitted when a TaxRate is supplied as input, such as in <a href="ctp:api:type:OrderImportDraft">OrderImportDraft</a>, and then defaults to <code>Net</code>.</p>
+     *
+
+     * @return null|string
+     */
+    public function getTaxRoundingTarget();
+
+    /**
      * @param ?string $id
      */
     public function setId(?string $id): void;
@@ -130,4 +140,9 @@ interface TaxRate extends JsonObject
      * @param ?SubRateCollection $subRates
      */
     public function setSubRates(?SubRateCollection $subRates): void;
+
+    /**
+     * @param ?string $taxRoundingTarget
+     */
+    public function setTaxRoundingTarget(?string $taxRoundingTarget): void;
 }